+++ src/main/kotlin/com/github/victorrentea/livecoding/Extensions.kt
@@ -20,3 +20,8 @@ fun PsiReferenceExpression.isRead(): Boolean {
     return true
 }
 
+@Suppress("UNCHECKED_CAST")
+inline fun <reified T : Any> List<*>.checkItemsAre() =
+    if (all { it is T })
+        this as List<T>
+    else null