+++ src/main/kotlin/com/github/victorrentea/livecoding/lombok/AddRequiredArgsConstructorInspection.kt
@@ -1,19 +1,18 @@
 package com.github.victorrentea.livecoding.lombok
 
-import com.github.victorrentea.livecoding.lombok.AddRequiredArgsConstructorInspection.Companion.INSPECTION_NAME
-import com.github.victorrentea.livecoding.FrameworkDetector.lombokIsPresent
+import com.github.victorrentea.livecoding.FrameworkDetector
 import com.intellij.codeInspection.LocalInspectionTool
 import com.intellij.codeInspection.LocalQuickFixOnPsiElement
 import com.intellij.codeInspection.ProblemHighlightType
 import com.intellij.codeInspection.ProblemsHolder
 import com.intellij.openapi.command.WriteCommandAction
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.util.NlsSafe
 import com.intellij.openapi.util.TextRange
 import com.intellij.psi.*
 import com.intellij.psi.PsiModifier.FINAL
 import com.intellij.psi.PsiModifier.STATIC
 import com.intellij.psi.codeStyle.JavaCodeStyleManager
+import com.intellij.util.PsiErrorElementUtil
 import kotlin.math.min
 
 
@@ -22,9 +21,9 @@ class AddRequiredArgsConstructorInspection : LocalInspectionTool() {
         const val INSPECTION_NAME = "Final fields can be injected via @RequiredArgsConstructor"
         const val FIX_NAME = "Add @RequiredArgsConstructor (lombok)"
     }
+
     override fun buildVisitor(holder: ProblemsHolder, isOnTheFly: Boolean): PsiElementVisitor {
-        if (!lombokIsPresent(holder.file)) {
-//            println("NO LOMBOK")
+        if (!FrameworkDetector.lombokIsPresent(holder.file)) {
             return PsiElementVisitor.EMPTY_VISITOR
         }
         return AddRequiredArgsConstructorVisitor(holder)
@@ -38,13 +37,7 @@ class AddRequiredArgsConstructorInspection : LocalInspectionTool() {
             if (!field.hasModifierProperty(FINAL)) return
             if (field.hasInitializer()) return // KO for private final int x = 2
             val psiClass = field.containingClass ?: return
-            if (psiClass.constructors.size >= 2) return
-
-            val existingConstructor = psiClass.constructors.getOrNull(0)
-            if (existingConstructor != null) {
-                if (!constructorOnlyCopiesParamsToFields(existingConstructor)) return
-                if (existingConstructor.parameterList.parameters.map{ it.name }.contains(field.name)) return
-            }
+            if (psiClass.constructors.isNotEmpty()) return
 
             val textLength = min(
                 field.nameIdentifier.textRangeInParent.endOffset + 1,