+++ src/test/live-coding-playground/src/main/java/lombok/replacerac/ReplaceRAC_TwoArg_withoutThis.java
@@ -1,0 +1,11 @@
+package lombok.replacerac;
+
+public class ReplaceRAC_TwoArg_withoutThis {
+    private final int dummy;
+    private final int dummy2;
+
+    public ReplaceRAC_TwoArg_withoutThis(int dummy_, int dummy2_) { //
+        dummy = dummy_;
+        dummy2 = dummy2_;
+    }
+}