+++ src/test/live-coding-playground/src/test/java/lombok/replacerac/ReplaceRAC_TwoArg_withoutThis.java
@@ -1,0 +1,9 @@
+package lombok.replacerac;
+
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public class ReplaceRAC_TwoArg_withoutThis {
+    private final int dummy;
+    private final int dummy2;
+}